% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRda.R
\name{checkRda}
\alias{checkRda}
\title{Check which objects are stored in a given .rda file}
\usage{
checkRda(file)
}
\arguments{
\item{file}{The .rda file to check}
}
\value{
A vector containing the names of objects in the .rda file.
}
\examples{
TempFile <- tempfile(fileext = ".rda") 
temp1 <- 10; temp2 <- c(20, 30);
save(temp1, temp2, file = TempFile)
checkRda(TempFile)
unlink(TempFile); rm(temp1, temp2, TempFile)
}
\author{
Sam Vennell <Sam.Vennell@mbie.govt.nz> and Eric Wu <eric.wu@mbie.govt.nz>
}

